Extract mouse UI pointer target resolution
Move pointer data construction, hierarchy and path lookup, drop target validation, and their result DTOs into a stateless resolver. Keep the existing out-parameter contract for move-only review, while removing the redundant RaycastUI wrapper in favor of direct UiRaycastHelper calls.
